Refactor CNC_Functions; Finish Move
Added Motion, Probe, Report, and Spindle to each handle specific aspects.

Delete CNC_Functions.h as no longer needed.

This code compiles and saves us 1% in program space and 1% in dynamic RAM.
However, I am sure there are bugs I need to discover.
