+++ cnc_ctrl_v1/Report.cpp
@@ -1,0 +1,61 @@
+/*This file is part of the Maslow Control Software.
+
+The Maslow Control Software is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Maslow Control Software is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with the Maslow Control Software.  If not, see <http://www.gnu.org/licenses/>.
+
+Copyright 2014-2017 Bar Smith*/
+
+// This file contains the functions for outgoing Serial responses
+
+#include "Maslow.h"
+
+void  returnError(){
+    /*
+    Prints the machine's positional error and the amount of space available in the 
+    gcode buffer
+    */
+        Serial.print(F("[PE:"));
+        Serial.print(leftAxis.error());
+        Serial.print(',');
+        Serial.print(rightAxis.error());
+        Serial.print(',');
+        Serial.print(incSerialBuffer.spaceAvailable());
+        Serial.println(F("]"));
+}
+
+void  returnPoz(const float& x, const float& y, const float& z){
+    /*
+    Causes the machine's position (x,y) to be sent over the serial connection updated on the UI
+    in Ground Control. Only executes if hasn't been called in at least timeout ms.
+    */
+    
+    static unsigned long lastRan = millis();
+    unsigned int         timeout = 200;
+    
+    if (millis() - lastRan > timeout){
+        
+        
+        Serial.print(F("<Idle,MPos:"));
+        Serial.print(x/sys.inchesToMMConversion);
+        Serial.print(F(","));
+        Serial.print(y/sys.inchesToMMConversion);
+        Serial.print(F(","));
+        Serial.print(z/sys.inchesToMMConversion);
+        Serial.println(F(",WPos:0.000,0.000,0.000>"));
+        
+        returnError();
+        
+        lastRan = millis();
+    }
+    
+}
